DTSP-5154: moved toggle logic
