+++ src/main/java/uk/gov/hmcts/probate/service/FeatureToggleService.java
@@ -23,9 +23,6 @@ public class FeatureToggleService {
     private static final String DECLARATION_NOT_SIGNED_TOGGLE = "probate-cron-declaration-not-signed";
     private static final String NFI_DATA_EXTRACT_TOGGLE = "probate-nfi-data-extract";
 
-    @Value("${probate.wa.enabled}")
-    private boolean probateWAEnabled;
-
     @Autowired
     public FeatureToggleService(LDClient ldClient, @Value("${ld.user.key}") String ldUserKey,
                                 @Value("${ld.user.firstName}") String ldUserFirstName,
@@ -64,10 +61,6 @@ public class FeatureToggleService {
     public boolean isPoundValueFeatureToggleOn() {
         return this.isFeatureToggleOn(
                 SMEE_AND_FORD_POUND_VALUE_TOGGLE, false);
-    }
-
-    public boolean isProbateWAEnabledToggleOn() {
-        return probateWAEnabled;
     }
 
     public boolean isIronMountainInBackOffice() {
